working on long click for portfolio list

+++ MonoStockPortfolio.Entities/Portfolio.cs
@@ -7,7 +7,8 @@ namespace MonoStockPortfolio.Entities
     public class Portfolio
     {
         public Portfolio() { }
-        public Portfolio(int id) { ID = id; }
+        public Portfolio(long id) { ID = id; }
+
         public long? ID { get; private set; }
         public string Name { get; set; }
